updated integers format

+++ sdk/go/server/model_cost.go
@@ -17,7 +17,7 @@ package server
 type Cost struct {
 
 	// The unit price information.
-	Price int32 `json:"price"`
+	Price float32 `json:"price"`
 
 	// The associated currency.
 	Currency string `json:"currency"`
